# Conflicts:
#	buildconfig/dependencies.gradle
#	tkpdcore/src/main/java/com/tokopedia/core/base/di/scope/ActivityScope.java
#	tkpdcore/src/main/java/com/tokopedia/core/database/DbFlowDatabase.java
#	tkpdcore/src/main/java/com/tokopedia/core/network/constants/TkpdBaseURL.java
#	tkpdcore/src/main/res/drawable-hdpi/chevron_right.png
#	tkpdcore/src/main/res/drawable-mdpi/chevron_right.png
#	tkpdcore/src/main/res/drawable-xhdpi/chevron_right.png
#	tkpdcore/src/main/res/drawable-xxhdpi/chevron_right.png
#	tkpdcore/src/main/res/drawable-xxxhdpi/chevron_right.png
#	tkpdseller/build.gradle
#	tkpdseller/src/main/AndroidManifest.xml
#	tkpdseller/src/main/java/com/tokopedia/seller/topads/view/presenter/TopAdsEtalaseListPresenterImpl.java
#	tkpdseller/src/main/java/com/tokopedia/seller/topads/view/widget/PrefixEditText.java
#	tkpdseller/src/main/res/values/colors.xml
#	tkpdseller/src/main/res/values/dimens.xml
#	tkpdseller/src/main/res/values/integers.xml
#	tkpdseller/src/main/res/values/strings.xml
#	tkpdseller/src/main/res/values/styles.xml

+++ customerapp/src/main/java/com/tokopedia/tkpd/home/recharge/util/CategoryComparator.java
@@ -1,7 +1,6 @@
 package com.tokopedia.tkpd.home.recharge.util;
 
 
-
 import com.tokopedia.core.database.model.category.Category;
 
 import java.util.Comparator;
@@ -14,7 +13,7 @@ public class CategoryComparator implements Comparator<Category> {
 
     @Override
     public int compare(Category c1, Category c2) {
-        return c1.getAttributes().getWeight().compareTo(c2.getAttributes().getWeight());
+        return ((Integer) c1.getAttributes().getWeight()).compareTo(c2.getAttributes().getWeight());
     }
 
 }
