+++ customerapp/src/main/java/com/tokopedia/tkpd/home/recharge/util/OperatorComparator.java
@@ -12,7 +12,7 @@ public class OperatorComparator implements Comparator<RechargeOperatorModel> {
 
     @Override
     public int compare(RechargeOperatorModel o1, RechargeOperatorModel o2) {
-        return o1.weight.compareTo(o2.weight);
+        return ((Integer) o1.weight).compareTo(o2.weight);
     }
 
 }
